Move music-related unit tests into a separate file
