+++ test/unit/blocks_sounds.js
@@ -1,18 +1,8 @@
 const test = require('tap').test;
 const Sound = require('../../src/blocks/scratch3_sound');
 let playedSound;
-let playedDrum;
-let playedInstrument;
-const runtime = {
-    audioEngine: {
-        numDrums: 3,
-        numInstruments: 3,
-        instrumentPlayer: {
-            loadInstrument: instrument => (playedInstrument = instrument)
-        }
-    }
-};
-const blocks = new Sound(runtime);
+
+const blocks = new Sound();
 const util = {
     target: {
         sprite: {
@@ -24,8 +14,7 @@ const util = {
             ]
         },
         audioPlayer: {
-            playSound: soundId => (playedSound = soundId),
-            playDrumForBeats: drum => (playedDrum = drum)
+            playSound: soundId => (playedSound = soundId)
         }
     }
 };
@@ -80,32 +69,5 @@ test('playSound prioritizes sound name if given a string', t => {
     blocks.playSound(args, util);
     // Use the sound named '6', which is the fourth
     t.strictEqual(playedSound, 'fourth soundId');
-    t.end();
-});
-
-test('playDrum uses 1-indexing and wrap clamps', t => {
-    let args = {DRUM: 1};
-    blocks.playDrumForBeats(args, util);
-    t.strictEqual(playedDrum, 0);
-
-    args = {DRUM: runtime.audioEngine.numDrums + 1};
-    blocks.playDrumForBeats(args, util);
-    t.strictEqual(playedDrum, 0);
-
-    t.end();
-});
-
-test('setInstrument uses 1-indexing and wrap clamps', t => {
-    // Stub getSoundState
-    blocks._getSoundState = () => ({});
-
-    let args = {INSTRUMENT: 1};
-    blocks.setInstrument(args, util);
-    t.strictEqual(playedInstrument, 0);
-
-    args = {INSTRUMENT: runtime.audioEngine.numInstruments + 1};
-    blocks.setInstrument(args, util);
-    t.strictEqual(playedInstrument, 0);
-
     t.end();
 });
